+++ scripts/pre-commit.sh
@@ -2,7 +2,7 @@
 
 # Prevent committing badly formatted code
 
-cargo fmt -- --check
+cargo +nightly-2024-03-12 fmt -- --check
 if [ $? -ne 0 ]; then
 	echo "Run \`cargo fmt\` to fix formatting issues before committing."
 	exit 1